requirements.txt
python-pptx
python-docx
google-generativeai
openai

